pyproject.toml
[project]
name = "making-apps-intelligent"
version = "0.1.0"
description = "Workshop materials for building MCP clients and intelligent applications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.39.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.1",
]
